+++ tests/testthat/test-sl_read.R
@@ -1,6 +1,6 @@
 testthat::test_that("reading in a file works", {
-  file = system.file("extdata/P30_wrist100.csv.gz", package = "stepcount")
-  if (asleep_check()) {
+  file = system.file("extdata/example_sleep.csv.gz", package = "asleep")
+  if (suppressWarnings(asleep_check())) {
     out = sl_read(file)
     testthat::expect_true(is.list(out))
     testthat::expect_named(out, c("data", "info"))