+++ NEWS.md
@@ -1,3 +1,4 @@
 # asleep 0.1.0
 
 * Initial submission.
+* Added all the functions and tests.